# swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Gerdt_LR1",
    "version": "1.0"
  },
  "paths": {
    "/api/Account/register": {
      "post": {
        "tags": [
          "Account"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Account": {
      "post": {
        "tags": [
          "Account"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginData"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginData"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginData"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Terms/show-term/{id}": {
      "get": {
        "tags": [
          "Terms"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Term"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Term"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Term"
                }
              }
            }
          }
        }
      }
    },
    "/api/Terms/change-term/{id}": {
      "put": {
        "tags": [
          "Terms"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Term"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Term"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Term"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Terms/create": {
      "post": {
        "tags": [
          "Terms"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Term"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Term"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Term"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Term"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Term"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Term"
                }
              }
            }
          }
        }
      }
    },
    "/api/Terms/{id}": {
      "delete": {
        "tags": [
          "Terms"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Terms/translate": {
      "post": {
        "tags": [
          "Terms"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranslateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TranslateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TranslateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Direction": {
        "enum": [
          "EnToRu",
          "RuToEn"
        ],
        "type": "string"
      },
...
      "LoginData": {
        "type": "object",
        "properties": {
          "login": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RegisterDto": {
        "required": [
          "login",
          "password"
        ],
        "type": "object",
        "properties": {
          "login": {
            "maxLength": 64,
            "minLength": 3,
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 3,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "Term": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "en": {
            "type": "string",
            "nullable": true
          },
          "ru": {
            "type": "string",
            "nullable": true
          },
          "domain": {
            "$ref": "#/components/schemas/TermDomain"
          }
        },
        "additionalProperties": false
      },
      "TermDomain": {
        "enum": [
          "General",
          "Drilling",
          "Geology",
          "Equipment",
          "Safety"
        ],
        "type": "string"
      },
      "TranslateDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "nullable": true
          },
          "direction": {
            "$ref": "#/components/schemas/Direction"
          }
        },
        "additionalProperties": false
      }
    }
  }
}